+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 3.0.1
 
-###### _April 22 2023_
+###### _May 13 2023_
 
 - UNDER DEVELOPMENT